+++ src/common/js/functions.js
@@ -3,7 +3,14 @@
  */
 
 import _ from 'lodash';
-const assert = require( 'assert' ).strict; // up to nodejs v16.x
+import { strict as assert } from 'node:assert';
+
+import { Tracker } from 'meteor/tracker';
+
+AccountsHub._instances = {
+    dep: new Tracker.Dependency(),
+    data: {}
+};
 
 /**
  * @locus Anywhere
@@ -23,4 +30,44 @@ AccountsHub.areSame = function( userA, userB ){
     }
     const res = ( idA === idB );
     return res;
+}
+
+/**
+ * @locus Anywhere
+ * @param {String} name the name of a tabular
+ * @returns {ahClass} the corresponding ahClass instance, or null
+ */
+AccountsHub.getByTabularName = function( name ){
+    _trace( 'AccountsHub.getByTabularName()', arguments );
+    assert( name && _.isString( name ), 'expects a string, got '+name );
+    let found = null;
+    Object.values( AccountsHub._instances.data ).every(( it ) => {
+        if( it.tabularName() === name ){
+            found = it;
+        }
+        return !found;
+    });
+    return found;
+}
+
+/**
+ * Getter/Setter
+ * @locus Anywhere
+ * @param {String} name
+ * @param {ahClass} instance
+ * @returns {ahClass} the named ahClass instance, or null
+ *  A reactive data source
+ */
+AccountsHub.getInstance = function( name, instance ){
+    _trace( 'AccountsHub.getInstance()', arguments );
+    assert( name && _.isString( name ), 'expects a string, got '+name );
+    if( instance ){
+        assert( instance instanceof AccountsHub.ahClass, 'expects an instance of AccountsHub.ahClass, got '+instance );
+        AccountsHub._instances.data[name] = instance;
+        AccountsHub._instances.dep.changed();
+    } else {
+        instance = AccountsHub._instances.data[name] || null;
+        AccountsHub._instances.dep.depend();
+    }
+    return instance;
 }