+++ src/common/js/global.js
@@ -3,5 +3,4 @@
  */
 
 AccountsHub = {
-    instances: {}
 };